auth done

+++ controllers/feed.js
@@ -1,6 +1,7 @@
 const { validationResult } = require("express-validator");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 const { clearImage } = require("../util/image");
 
@@ -57,6 +58,7 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const errors = validationResult(req);
+  let creator;
   if (!errors.isEmpty()) {
     const error = new Error("Validation failed on server side");
     error.status = 422;
@@ -72,16 +74,67 @@ exports.createPost = (req, res, next) => {
     title,
     content,
     imageUrl: imageUrl,
-    creator: {
-      name: "Teddy",
-    },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
-      res.status(201).json({
-        message: "post created successfully",
-        post: result,
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
+      return res.status(201).json({
+        message: "Post created successfully",
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
+      });
+    })
+    .catch((err) => {
+      if (!err.status) {
+        err.status = 500;
+      }
+      next(err);
+    });
+};
+
+exports.getStatus = (req, res, next) => {
+  const userId = req.userId;
+  User.findById(userId)
+    .then((user) => {
+      if (!user) {
+        throw new Error("Could not retrieve user");
+      }
+      return res.status(200).json({
+        status: user.status,
+      });
+    })
+    .catch((err) => {
+      if (!err.status) {
+        err.status = 500;
+      }
+      next(err);
+    });
+};
+
+exports.updateStatus = (req, res, next) => {
+  const userId = req.userId;
+  const status = req.body.status;
+  User.findById(userId)
+    .then((user) => {
+      if (!user) {
+        throw new Error("Could not retrieve user");
+      }
+      console.log(user);
+      user.status = status;
+      return user.save();
+    })
+    .then((user) => {
+      return res.status(200).json({
+        status: user.status,
       });
     })
     .catch((err) => {
@@ -118,6 +171,11 @@ exports.updatePost = (req, res, next) => {
         err.status = 404;
         throw err;
       }
+      if (post.creator.toString() !== req.userId.toString()) {
+        const err = new Error("Not Authorized");
+        err.status = 403;
+        throw err;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -149,12 +207,22 @@ exports.deletePost = (req, res, next) => {
         err.status = 404;
         throw err;
       }
-      // Check logged in user
+      if (post.creator.toString() !== req.userId.toString()) {
+        const err = new Error("Not Authorized");
+        err.status = 403;
+        throw err;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(post._id);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save;
+    })
+    .then((result) => {
       res.status(200).json({
         message: "Post deleted successfully",
       });
